api: sched: config options for event accumulation
Currently ODP assumes that scheduler implementation will use best
parameters for burst accumulation size and timeout. However HW
schedulers might require this to be setup in control path and not on
each call of odp_schedule_multi().

Exposing such parameters to the application allows for better control of
scheduler latency and throughput as well as benefit from more effective
SIMD processing.

This patch proposes the following configuration options to configure the
multi-event aspects of the scheduler API.

1. Number of events to accumulate
The scheduler will asynchronously accumulate upto these
many number of events and deliver the events on the subsequent
invocation of odp_schedule_multi().

Invoking odp_schedule_multi() with burst less than this value is
not allowed.

2. Timeout for accumulation:
The scheduler will asynchronously accumulate events for this
time period before delivering the batch of events.

Signed-off-by: Vidya Velumuri <vvelumuri@marvell.com>